Move sample_times method
