+++ gammapy/time/__init__.py
@@ -1,5 +1,4 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Time analysis."""
 from .period import *
-from .simulate import *
 from .variability import *
