+++ gammapy/time/tests/test_variability.py
@@ -4,7 +4,8 @@ import numpy as np
 from numpy.testing import assert_allclose
 from astropy.table import Column, Table
 from astropy.time import Time
-from gammapy.time import LightCurve, compute_chisq, compute_fvar
+from gammapy.estimators import LightCurve
+from gammapy.time import compute_chisq, compute_fvar
 from gammapy.utils.testing import assert_quantity_allclose
